backend/pyproject.toml
[tool.poetry]
name = "chatbetsite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
daphne = "^4.1.2"
django = "^5.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
channels-redis = "^4.2.0"
factory-boy = "^3.3.1"
django-polymorphic = "^3.1.0"
django-rest-polymorphic = "^0.1.10"
scrapy = "^2.11.2"
django-selenium = "^0.9.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
